code reorganized in view for readability and checkstyle and create inventory moved out of authentication package
