+++ src/main/java/app/AppBuilder.java
@@ -19,8 +19,6 @@ import entity.UserFactory;
 import interface_adapter.ViewManagerModel;
 import interface_adapter.add_to_inventory.AddToInventoryController;
 import interface_adapter.add_to_inventory.AddToInventoryPresenter;
-import interface_adapter.adoption.AdoptionController;
-import interface_adapter.adoption.AdoptionPresenter;
 import interface_adapter.adoption.AdoptionViewModel;
 import interface_adapter.cat_image.CatImageController;
 import interface_adapter.cat_image.CatImagePresenter;
@@ -48,9 +46,9 @@ import interface_adapter.timer.TimerPresenter;
 import interface_adapter.timer.TimerViewModel;
 import interface_adapter.use_item_in_inventory.UseItemController;
 import interface_adapter.use_item_in_inventory.UseItemPresenter;
-import use_case.authentication.create_inventory.CreateInventoryInputBoundary;
-import use_case.authentication.create_inventory.CreateInventoryInteractor;
-import use_case.authentication.create_inventory.CreateInventoryOutputBoundary;
+import use_case.create_inventory.CreateInventoryInputBoundary;
+import use_case.create_inventory.CreateInventoryInteractor;
+import use_case.create_inventory.CreateInventoryOutputBoundary;
 import use_case.display_cat_image.CatImageDataAccessInterface;
 // import use_case.cat_image.CatImageInputBoundary;
 // import use_case.cat_image.CatImageInteractor;
@@ -58,7 +56,6 @@ import use_case.display_cat_image.CatImageDataAccessInterface;
 import use_case.change_password.ChangePasswordInputBoundary;
 import use_case.change_password.ChangePasswordInteractor;
 import use_case.change_password.ChangePasswordOutputBoundary;
-import use_case.display_cat_image.CatImageDataAccessInterface;
 import use_case.display_cat_image.CatImageInputBoundary;
 import use_case.display_cat_image.CatImageInteractor;
 import use_case.display_cat_image.CatImageOutputBoundary;
