+++ src/main/java/app/InventoryExample.java
@@ -15,7 +15,10 @@ import interface_adapter.create_inventory.CreateInventoryPresenter;
 import interface_adapter.create_inventory.InventoryViewModel;
 import interface_adapter.use_item_in_inventory.UseItemController;
 import interface_adapter.use_item_in_inventory.UseItemPresenter;
-import use_case.authentication.create_inventory.*;
+import use_case.create_inventory.CreateInventoryInputBoundary;
+import use_case.create_inventory.CreateInventoryInteractor;
+import use_case.create_inventory.CreateInventoryInventoryDataAccessInterface;
+import use_case.create_inventory.CreateInventoryOutputBoundary;
 import use_case.food_management.add_to_inventory.AddToInventoryDataAccessInterface;
 import use_case.food_management.add_to_inventory.AddToInventoryInputBoundary;
 import use_case.food_management.add_to_inventory.AddToInventoryInteractor;
