+++ src/main/java/use_case/create_inventory/CreateInventoryInputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.authentication.create_inventory;
+package use_case.create_inventory;
 
 /**
  * Input boundry of the Create Inventory Use Case.
